move `filterDataPerClauses` to core
