+++ packages/core/package.json
@@ -20,7 +20,9 @@
   "dependencies": {
     "is-what": "^3.14.1",
     "merge-anything": "^4.0.2",
-    "ts-toolbelt": "^9.6.0"
+    "ts-toolbelt": "^9.6.0",
+    "fast-sort": "^3.1.3",
+    "path-to-prop": "^2.0.2"
   },
   "devDependencies": {
     "@magnetarjs/test-utils": "workspace:*"
