+++ packages/core/src/helpers/parseValueForFilters.ts
