+++ packages/core/src/index.ts
@@ -2,6 +2,8 @@ export * from './Magnetar'
 export { CollectionInstance } from './Collection'
 export { DocInstance } from './Doc'
 export * from './helpers/pathHelpers'
+export * from './helpers/dataHelpers'
+export * from './helpers/parseValueForFilters'
 export * from './types/plugins'
 export * from './types/actions'
 export * from './types/atoms'
