+++ packages/plugin-vue2/src/helpers/dataHelpers.ts
@@ -1,86 +1,3 @@
-import { Clauses } from '@magnetarjs/core'
-import { isNumber, isArray } from 'is-what'
-import { getProp } from 'path-to-prop'
-import { sort, ISortByObjectSorter } from 'fast-sort'
-import { parseValueForFilters } from '@magnetarjs/utils'
-
-/**
- * Filters a Collection module's data based on provided clauses.
- *
- * @param {Record<string, Record<string, any>>} collectionDB
- * @param {Clauses} clauses
- * @returns {'no-filter' | [string, Record<string, any>][]} filtered and sorted entries
- */
-export function filterDataPerClauses(
-  collectionDB: Record<string, Record<string, any>>,
-  clauses: Clauses
-): 'no-filter' | [string, Record<string, any>][] {
-  const { where = [], orderBy = [], limit } = clauses
-  // return 'no-filter' if no filtering is needed
-  if (!where.length && !orderBy.length && !isNumber(limit)) return 'no-filter'
-  // all other cases we need to create a new Map() with the results
-  const entries: [string, Record<string, any>][] = []
-  Object.entries(collectionDB).forEach(([docId, docData]) => {
-    const passesWhereFilters = where.every((whereQuery) => {
-      const [fieldPath, operator, expectedValue] = whereQuery
-      const valueAtFieldPath = parseValueForFilters(getProp(docData, fieldPath) as any)
-      let passes = false
-      switch (operator) {
-        case '==':
-          passes = valueAtFieldPath == expectedValue
-          break
-        case '!=':
-          passes = valueAtFieldPath != expectedValue
-          break
-        case '<':
-          passes = valueAtFieldPath < expectedValue
-          break
-        case '<=':
-          passes = valueAtFieldPath <= expectedValue
-          break
-        case '>':
-          passes = valueAtFieldPath > expectedValue
-          break
-        case '>=':
-          passes = valueAtFieldPath >= expectedValue
-          break
-        case 'in':
-          passes = isArray(expectedValue) && expectedValue.includes(valueAtFieldPath)
-          break
-        case 'not-in':
-          passes = isArray(expectedValue) && !expectedValue.includes(valueAtFieldPath)
-          break
-        case 'array-contains':
-          passes = isArray(valueAtFieldPath) && valueAtFieldPath.includes(expectedValue)
-          break
-        case 'array-contains-any':
-          passes =
-            isArray(valueAtFieldPath) &&
-            valueAtFieldPath.some((v: any) => isArray(expectedValue) && expectedValue.includes(v))
-          break
-        default:
-          throw new Error('invalid operator')
-      }
-      return passes
-    })
-    if (!passesWhereFilters) return
-    entries.push([docId, docData])
-  })
-  // orderBy
-  const by = orderBy.reduce((carry, [path, direction = 'asc']) => {
-    const sorter: ISortByObjectSorter<[string, Record<string, any>]> = {
-      [direction as 'asc']: (entry: [string, Record<string, any>]) => getProp(entry[1], path),
-    }
-    carry.push(sorter)
-    return carry
-  }, [] as ISortByObjectSorter<[string, Record<string, any>]>[])
-  const entriesOrdered = orderBy.length ? sort(entries).by(by) : entries
-  // limit
-  const entriesLimited = isNumber(limit) ? entriesOrdered.slice(0, limit) : entriesOrdered
-
-  return entriesLimited
-}
-
 type CustomMap<DocDataType = Record<string, any>> = {
   // data
   /**
