+++ packages/plugin-vue3/package.json
@@ -24,7 +24,6 @@
     "@magnetarjs/core": "workspace:*",
     "@magnetarjs/utils": "workspace:*",
     "copy-anything": "^2.0.3",
-    "fast-sort": "^3.1.1",
     "filter-anything": "^2.2.1",
     "is-what": "^3.14.1",
     "merge-anything": "^4.0.2",
