+++ packages/plugin-vue3/test/internal/_vueComputed.ts
