+++ packages/test-utils/src/PluginRemoteMock/actions.ts
@@ -19,10 +19,10 @@ import {
   PluginInsertActionPayload,
   PluginWriteActionPayload,
   Clauses,
+  filterDataPerClauses,
 } from '../../../core/src'
 import { StorePluginModuleConfig, RemoteStoreOptions } from './index'
 import { throwIfEmulatedError, waitMs, pokedexMap, generateRandomId } from '../helpers'
-import { filterDataPerClauses } from './helpers'
 import { isFullArray, isPromise } from 'is-what'
 
 export function writeActionFactory(
@@ -105,13 +105,13 @@ function mockDataRetrieval(
   pluginModuleConfig: StorePluginModuleConfig
 ): Record<string, any>[] {
   if (docId === 'trainer') return [{ name: 'Luca', age: 10, dream: 'job' }]
-  
+
   if (collectionPath === 'pokedex') {
     const _pokedexMap = pokedexMap()
-    
+
     if (docId) {
       const result = _pokedexMap.get(docId)
-      return (result) ? [result] : []
+      return result ? [result] : []
     }
 
     const clauses: Clauses = pick(pluginModuleConfig, ['where', 'orderBy', 'limit'])
@@ -175,17 +175,19 @@ export function streamActionFactory(storePluginOptions: RemoteStoreOptions): Plu
     // this mocks actual data coming in at different intervals
     dataRetrieved.forEach((data, i) => {
       const metaData = { data, id: data.id || docId, exists: true }
-      
+
       // we can simulate new docs coming in by manually triggerering promises
       if (isFullArray(payload) && payload.every(isPromise)) {
         // the payload is an array full of promises!
         // in this case we will emit every time the promises at the index is triggered
         const promise = payload[i]
         if (!promise) return
-        promise.then(() => { mustExecuteOnRead.added(data, metaData) })
+        promise.then(() => {
+          mustExecuteOnRead.added(data, metaData)
+        })
         return
       }
-      
+
       // otherwise emulate server response with setTimeout (can be wonky in automated tests)
       const waitTime = 10 + i * 200
       setTimeout(() => {
