+++ packages/utils/package.json
@@ -18,6 +18,12 @@
   "dependencies": {
     "is-what": "^3.14.1"
   },
+  "peerDependencies": {
+    "@magnetarjs/core": "workspace:*"
+  },
+  "devDependencies": {
+    "@magnetarjs/core": "workspace:*"
+  },
   "keywords": [
     "magnetar",
     "magnetarjs"
